Début d'implémentation des permissions et gamerule. Ca permettra de rendre personnalisable plus facilement son mode de jeu sur Minecraft

+++ api/src/main/java/fr/euphyllia/skyfolia/api/utils/BitwiseFlag.java
@@ -2,13 +2,13 @@ package fr.euphyllia.skyfolia.api.utils;
 
 public abstract class BitwiseFlag {
 
-    protected int flags = 0;
+    protected long flags = 0;
 
-    protected boolean isFlagSet(int flag) {
+    protected boolean isFlagSet(long flag) {
         return ((this.flags & flag) == flag);
     }
 
-    protected void setFlags(int flag, boolean value) {
+    protected void setFlags(long flag, boolean value) {
         if (value) {
             this.flags |= flag;
         } else {
