+++ plugin/src/main/java/fr/euphyllia/skyfolia/managers/skyblock/PermissionManager.java
@@ -6,23 +6,23 @@ import fr.euphyllia.skyfolia.api.utils.BitwiseFlag;
 
 public class PermissionManager extends BitwiseFlag {
 
-    public PermissionManager(int actualFlags) {
+    public PermissionManager(long actualFlags) {
         this.flags = actualFlags;
     }
 
-    public static int valueOf(String permissions) {
+    public static long valueOf(String permissions) {
         return PermissionsIsland.permissionValue(permissions);
     }
 
-    public int getPermissions() {
+    public long getPermissions() {
         return this.flags;
     }
 
-    public void definePermission(int permission, boolean value) {
+    public void definePermission(long permission, boolean value) {
         this.setFlags(permission, value);
     }
 
-    public boolean hasPermission(int permission) {
+    public boolean hasPermission(long permission) {
         return this.isFlagSet(permission);
     }
